stream_ml/requirements.txt
pyzmq
uvicorn
pandas
fastapi
mlflow